Moved prune functionality into its own file.
